Created  individual step files for each page and moved existing functionality
