+++ cypress/e2e/pages/agreementClosures/addBulkClosurePage.js
